Create Logistic_Regression_Model
Background

Credit risk poses a classification problem that’s inherently imbalanced. The reason is that healthy loans easily outnumber risky loans. For this Challenge, various techniques to train and evaluate models with imbalanced classes will be used. A dataset of historical lending activity from a peer-to-peer lending services company, to build a model that can identify the creditworthiness of borrowers will be used. 

The instructions for this Challenge are divided into the following subsections:

    Splitting the Data into Training and Testing Sets

    Creating Logistic Regression Model with the Original Data

    Predicting Logistic Regression Model with Resampled Training Data

    Writing Credit Risk Analysis Report