Module_12_Challenge_credit_risk_resampling.ipynb
Credit Risk Classification

Credit risk poses a classification problem that’s inherently imbalanced. This is because healthy loans easily outnumber risky loans. In this Challenge, we used various techniques to train and evaluate models with imbalanced classes. We used a dataset of historical lending activity from a peer-to-peer lending services company to build a model that can identify the creditworthiness of borrowers.
This challenge consists of the following subsections: Splitting the Data into Training and Testing Sets, Creating a Logistic Regression Model with the Original Data and Predicting a Logistic Regression Model with Resampled Training Data.